反正更新了许多东西，自己看吧(  最重要的应该是 渲染顺序更新 还有DeltaTime的Uint32改成Uint64了；虽然没啥用吧，没人会运行30多天

+++ PlantVsZombies/ParticleSystem/ParticleEmitter.cpp
@@ -30,7 +30,7 @@ void ParticleEmitter::SetAutoDestroyTime(int frames) {
         autoDestroyTime = config.lifetime + 15.0f;
     }
     else {
-        autoDestroyTime = static_cast<int>(frames);
+        autoDestroyTime = static_cast<float>(frames);
     }
     autoDestroyTimer = 0.0f;
 }
